Refactored NewsCards to utilize the latest entries automatically.
